+++ smart-experiences/smart-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/smart-common
 
+## 0.1.0
+
+### Minor Changes
+
+- [`0998f6fb413`](https://bitbucket.org/atlassian/atlassian-frontend/commits/0998f6fb413) - Add signal prop for aborting fetch
+
 ## 0.0.3
 
 ### Patch Changes